+++ sdk/resourcemanager/network/armnetwork/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Release History
 
+## 6.3.0 (2024-12-18)
+### Features Added
+
+- New function `*VirtualAppliancesClient.BeginGetBootDiagnosticLogs(context.Context, string, string, *VirtualAppliancesClientBeginGetBootDiagnosticLogsOptions) (*runtime.Poller[VirtualAppliancesClientGetBootDiagnosticLogsResponse], error)`
+- New function `*VirtualAppliancesClient.BeginReimage(context.Context, string, string, *VirtualAppliancesClientBeginReimageOptions) (*runtime.Poller[VirtualAppliancesClientReimageResponse], error)`
+- New struct `VirtualApplianceBootDiagnosticParameters`
+- New struct `VirtualApplianceInstanceID`
+
+
 ## 6.2.0 (2024-12-09)
 ### Features Added